Move MaybeParIter to rten-base
