+++ rten-base/src/iter.rs
@@ -99,3 +99,70 @@ impl<I: SplitIterator + Send> Producer for ParIter<I> {
         (Self(left_inner), Self(right_inner))
     }
 }
+
+/// Wrapper around either a serial or parallel iterator, returned by
+/// [`MaybeParIter::maybe_par_iter`].
+pub enum MaybeParallel<PI: ParallelIterator, SI: Iterator<Item = PI::Item>> {
+    Serial(SI),
+    Parallel(PI),
+}
+
+impl<PI: ParallelIterator, SI: Iterator<Item = PI::Item>> MaybeParallel<PI, SI> {
+    pub fn for_each<F: Fn(PI::Item) + Send + Sync>(self, f: F) {
+        match self {
+            MaybeParallel::Serial(iter) => iter.for_each(f),
+            MaybeParallel::Parallel(iter) => iter.for_each(f),
+        }
+    }
+}
+
+/// Trait which allows use of Rayon parallelism to be conditionally enabled.
+///
+/// See <https://crates.io/crates/rayon-cond> for a more full-featured alternative.
+pub trait MaybeParIter {
+    type Item;
+    type ParIter: ParallelIterator<Item = Self::Item>;
+    type Iter: Iterator<Item = Self::Item>;
+
+    /// Return an iterator which executes either in serial on the current
+    /// thread, or in parallel in a Rayon thread pool if `parallel` is true.
+    fn maybe_par_iter(self, parallel: bool) -> MaybeParallel<Self::ParIter, Self::Iter>;
+}
+
+impl<Item, I: rayon::iter::IntoParallelIterator<Item = Item> + IntoIterator<Item = Item>>
+    MaybeParIter for I
+{
+    type Item = Item;
+    type ParIter = I::Iter;
+    type Iter = I::IntoIter;
+
+    fn maybe_par_iter(self, parallel: bool) -> MaybeParallel<Self::ParIter, Self::Iter> {
+        if parallel {
+            MaybeParallel::Parallel(self.into_par_iter())
+        } else {
+            MaybeParallel::Serial(self.into_iter())
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::atomic::{AtomicU32, Ordering};
+
+    use super::MaybeParIter;
+
+    #[test]
+    fn test_maybe_par_iter() {
+        let count = AtomicU32::new(0);
+        (0..1000).maybe_par_iter(false).for_each(|_| {
+            count.fetch_add(1, Ordering::SeqCst);
+        });
+        assert_eq!(count.load(Ordering::SeqCst), 1000);
+
+        let count = AtomicU32::new(0);
+        (0..1000).maybe_par_iter(true).for_each(|_| {
+            count.fetch_add(1, Ordering::SeqCst);
+        });
+        assert_eq!(count.load(Ordering::SeqCst), 1000);
+    }
+}
