+++ src/gemm.rs
@@ -12,15 +12,13 @@ use std::ops::{Add, Mul, Range};
 
 use rayon::prelude::*;
 use rten_base::byte_cast::Pod;
-use rten_base::iter::range_chunks;
+use rten_base::iter::{range_chunks, MaybeParIter};
 use rten_base::num::Identities;
 use rten_tensor::prelude::*;
 use rten_tensor::{
     Alloc, AssumeInit, GlobalAlloc, Matrix, MatrixLayout, MatrixMut, NdTensor, NdTensorView,
     Storage,
 };
-
-use crate::iter_util::MaybeParIter;
 
 mod errors;
 mod im2col;
