+++ src/lib.rs
@@ -119,7 +119,6 @@ mod env;
 mod gemm;
 mod graph;
 mod header;
-mod iter_util;
 mod model;
 mod model_metadata;
 mod op_registry;
